requirements.txt
langchain
langchain-openai
langchain-pinecone
langchainhub
python-dotenv
